# Closes #2300 - `segString` Strip performance issue