@joshmarshall1 2023-04-19
> running both methods with an problem size larger than `10**4` produce an out of memory error

Added issue #2389 to track the OOM error